+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support non-string types in query and header parameters.
 - Support additionalProperties when no properties are defined, creating maps.
 
+### Fixed
+- Support references in responses.
+
 ## [0.0.1] - 2017-10-15
 
 ### Added